+write cache efforts +fixed bug in send/recv with done == 0 +trylock() in server's keep-alive instead of locked() +fixed ret value of handle_write() +answer.ret is now int32_t +cleanups +version changed to 0.8
git-svn-id: https://svn.code.sf.net/p/remotefs/code/trunk@10 7764f161-3e89-4957-9ae6-8e1f4830ae4a

+++ src/alloc.c
@@ -5,10 +5,10 @@
 #include "config.h"
 
 static char *last_allocated = NULL;
-static unsigned last_allocated_size = 0;
+static size_t last_allocated_size = 0;
 static int freed = 0;
 
-void *mp_alloc(unsigned size)
+void *mp_alloc(size_t size)
 {
 	if (freed != 0
 	&& last_allocated != NULL)
